Moved all common tree logic
