+++ lua/gitlab/actions/trees/init.lua
@@ -2,6 +2,7 @@ local u = require("gitlab.utils")
 local au = require("gitlab.actions.utils")
 local state = require("gitlab.state")
 local NuiTree = require("nui.tree")
+local NuiLine = require("nui.line")
 local M = {}
 
 local attach_uuid = function(str)
@@ -39,6 +40,7 @@ local function build_note_body(note, resolve_info)
 
   return noteHeader, text_nodes
 end
+
 ---Build note node
 ---@param note Note|DraftNote
 ---@param resolve_info table?
@@ -59,5 +61,59 @@ M.build_note = function(note, resolve_info)
 
   return note_node, text, text_nodes
 end
+
+---Inspired by default func https://github.com/MunifTanjim/nui.nvim/blob/main/lua/nui/tree/util.lua#L38
+M.nui_tree_prepare_node = function(node)
+  if not node.text then
+    error("missing node.text")
+  end
+
+  local texts = node.text
+  if type(node.text) ~= "table" or node.text.content then
+    texts = { node.text }
+  end
+
+  local lines = {}
+
+  for i, text in ipairs(texts) do
+    local line = NuiLine()
+
+    line:append(string.rep("  ", node._depth - 1))
+
+    if i == 1 and node:has_children() then
+      line:append(node:is_expanded() and " " or " ")
+      if node.icon then
+        line:append(node.icon .. " ", node.icon_hl)
+      end
+    else
+      line:append("  ")
+    end
+
+    line:append(text, node.text_hl)
+
+    local note_id = tostring(node.is_root and node.root_note_id or node.id)
+
+    local e = require("gitlab.emoji")
+
+    ---@type Emoji[]
+    local emojis = state.DISCUSSION_DATA.emojis[note_id]
+    local placed_emojis = {}
+    if emojis ~= nil then
+      for _, v in ipairs(emojis) do
+        local icon = e.emoji_map[v.name]
+        if icon ~= nil and not u.contains(placed_emojis, icon.moji) then
+          line:append(" ")
+          line:append(icon.moji)
+          table.insert(placed_emojis, icon.moji)
+        end
+      end
+    end
+
+    table.insert(lines, line)
+  end
+
+  return lines
+end
+
 
 return M
